+++ gotham-client/src/ecdsa/keygen.rs
@@ -15,12 +15,17 @@ use kms::chain_code::two_party as chain_code;
 use kms::ecdsa::two_party::*;
 use multi_party_ecdsa::protocols::two_party_ecdsa::lindell_2017::*;
 
-use super::super::api;
+use super::types::PrivateShare;
 use super::super::utilities::requests;
+use super::super::ClientShim;
+
+// iOS bindings
+use std::ffi::{CStr, CString};
+use std::os::raw::c_char;
 
 const KG_PATH_PRE: &str = "ecdsa/keygen";
 
-pub fn get_master_key(client_shim: &api::ClientShim) -> api::PrivateShare {
+pub fn get_master_key(client_shim: &ClientShim) -> PrivateShare {
     let start = PreciseTime::now();
 
     let res_body = requests::post(client_shim, &format!("{}/first", KG_PATH_PRE)).unwrap();
@@ -121,5 +126,36 @@ pub fn get_master_key(client_shim: &api::ClientShim) -> api::PrivateShare {
     let end = PreciseTime::now();
     println!("(id: {}) Took: {}", id, start.to(end));
 
-    api::PrivateShare { id, master_key }
+    PrivateShare { id, master_key }
+}
+
+#[no_mangle]
+pub extern "C" fn get_client_master_key(
+    c_endpoint: *const c_char,
+    c_auth_token: *const c_char,
+) -> *mut c_char {
+    let raw_endpoint = unsafe { CStr::from_ptr(c_endpoint) };
+    let endpoint = match raw_endpoint.to_str() {
+        Ok(s) => s,
+        Err(_) => panic!("Error while decoding raw endpoint"),
+    };
+
+    let raw_auth_token = unsafe { CStr::from_ptr(c_auth_token) };
+    let auth_token = match raw_auth_token.to_str() {
+        Ok(s) => s,
+        Err(_) => panic!("Error while decoding auth token"),
+    };
+
+    let client_shim = ClientShim::new(endpoint.to_string(), Some(auth_token.to_string()));
+
+    let private_share: PrivateShare = get_master_key(&client_shim);
+
+    let private_share_json = match serde_json::to_string(&private_share) {
+        Ok(share) => share,
+        Err(_) => panic!("Error while performing keygen to endpoint {}", endpoint),
+    };
+
+    CString::new(private_share_json.to_owned())
+        .unwrap()
+        .into_raw()
 }
