+++ gotham-client/src/ecdsa/mod.rs
@@ -9,3 +9,8 @@
 pub mod keygen;
 pub mod rotate;
 pub mod sign;
+pub mod types;
+
+pub use keygen::get_master_key;
+pub use rotate::rotate_master_key;
+pub use sign::sign;
