+++ gotham-client/src/ecdsa/types.rs
@@ -1,0 +1,7 @@
+use kms::ecdsa::two_party::MasterKey2;
+
+#[derive(Serialize, Deserialize)]
+pub struct PrivateShare {
+    pub id: String,
+    pub master_key: MasterKey2,
+}
